move new tests over
